+++ app/src/main/res/values/api_hosts.xml
@@ -40,7 +40,7 @@
         <item>https://vulcan.net.pl/?login</item>
         <item>https://vulcan.net.pl/?login</item>
         <item>https://vulcan.net.pl/?login</item>
-        <item>http://fakelog.tk/?email</item>
+        <item>http://fakelog.cf/?email</item>
     </string-array>
     <string-array name="hosts_symbols">
         <item>Default</item>
